requirements.txt
# Core
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.0
requests==2.31.0
apscheduler==3.10.4
pytz==2024.1
loguru==0.7.2
feedparser==6.0.11

# Market Data
kiteconnect==5.0.1

# Sentiment / NLP
transformers==4.38.0
torch==2.2.0
sentencepiece==0.2.0

# Local LLM via Ollama (OpenAI-compatible API)
openai==1.14.0

# Cloud LLM fallback
google-generativeai==0.5.0

# Agent Orchestration
langgraph==0.1.0
langchain-core==0.1.0

# Backtesting
backtrader==1.9.78.123

# Reinforcement Learning
stable-baselines3==2.2.1
gymnasium==0.29.1

# Database
SQLAlchemy==2.0.25

# API / Dashboard Backend
fastapi==0.110.0
uvicorn==0.27.0
websockets==12.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.0

# Visualization (reports)
matplotlib==3.8.0
plotly==5.18.0

